user-system/requirements.txt
# 用户系统依赖

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
mysql-connector-python==8.2.0

# 认证
pyjwt==2.8.0
bcrypt==4.1.2

# 数据验证
pydantic==2.5.0
pydantic[email]==2.5.0

# 环境配置
python-dotenv==1.0.0

# 生产环境（可选）
gunicorn==21.2.0

# 监控（可选）
prometheus-client==0.19.0

# 测试（可选）
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
